requirements.txt
pandas==0.24.2
chardet==3.0.4
numpy==1.16.2
DAWG==0.7.8
jieba==0.39
matplotlib==3.0.3
pypinyin==0.35.2
scikit_learn==0.20.3
tflearn==0.3.2
threadpool==1.3.2
tornado==6.0.2
websocket_client==0.56.0
tensorflow==1.8.0
netifaces==0.10.9
pymysql==0.8.0
